Rename TwigLinter to PurifierLinter
Since it does not lint twig templates, I created a more fitting name.
Moved PurifierLinterTest to unit tests
Renamed test data files to be more precise on what they contain.
