+++ bin/purifier_lint
@@ -9,4 +9,4 @@ if ( file_exists( $a = __DIR__.'/../../../autoload.php' ) ) {
 	require_once __DIR__.'/../vendor/autoload.php';
 }
 
-exit( ( new WMDE\Fundraising\ContentProvider\TwigLinter() )->lint( $_SERVER['argv'][1] ?? null ) );
+exit( ( new \WMDE\Fundraising\ContentProvider\Linter\PurifierLinter() )->lint( $_SERVER['argv'][1] ?? null ) );
