+++ test/Unit/PurifierLinterTest.php
@@ -2,15 +2,18 @@
 
 declare( strict_types=1 );
 
-namespace WMDE\Fundraising\ContentProvider\Test;
+namespace WMDE\Fundraising\ContentProvider\Test\Unit;
 
 use PHPUnit\Framework\TestCase;
-use WMDE\Fundraising\ContentProvider\TwigLinter;
+use WMDE\Fundraising\ContentProvider\Linter\PurifierLinter;
 
-class TwigLinterTest extends TestCase {
+/**
+ * @covers PurifierLinter
+ */
+class PurifierLinterTest extends TestCase {
 
 	public function testGivenNullFileName_errorIsOutput() {
-		$linter = new TwigLinter( function( $handle, string $errorMessage ) {
+		$linter = new PurifierLinter( function( $handle, string $errorMessage ) {
 			$this->assertSame( 'ERROR: Required file name argument not provided' . PHP_EOL, $errorMessage );
 		} );
 
@@ -18,7 +21,7 @@ class TwigLinterTest extends TestCase {
 	}
 
 	public function testGivenEmptyFileName_errorIsOutput() {
-		$linter = new TwigLinter( function( $handle, string $errorMessage ) {
+		$linter = new PurifierLinter( function( $handle, string $errorMessage ) {
 			$this->assertSame( 'ERROR: Required file name argument not provided' . PHP_EOL, $errorMessage );
 		} );
 
@@ -28,7 +31,7 @@ class TwigLinterTest extends TestCase {
 	public function testGivenNonExistingFileName_errorIsOutput() {
 		$errorWasOutput = false;
 
-		$linter = new TwigLinter( function( $handle, string $errorMessage ) use ( &$errorWasOutput ) {
+		$linter = new PurifierLinter( function( $handle, string $errorMessage ) use ( &$errorWasOutput ) {
 			$this->assertSame( 'ERROR: Could not read file' . PHP_EOL, $errorMessage );
 			$errorWasOutput = true;
 		} );
@@ -39,24 +42,24 @@ class TwigLinterTest extends TestCase {
 	}
 
 	public function testGivenNameOfFileWithValidHtml_noErrorsAreOutput() {
-		$linter = new TwigLinter( function( $handle, string $errorMessage ) {
+		$linter = new PurifierLinter( function( $handle, string $errorMessage ) {
 			$this->fail( 'No errors should be output, yet got: ' . $errorMessage );
 		} );
 
 		$this->expectOutputRegex( '/OK/' );
-		$this->assertSame( 0, $linter->lint( __DIR__ . '/data/ValidTwigFile.twig' ) );
+		$this->assertSame( 0, $linter->lint( __DIR__ . '/../data/ValidHtmlFile.twig' ) );
 	}
 
 	public function testGivenNameOfFileWithInvalidHtml_errorIsOutput() {
 		$errorWasOutput = false;
 
-		$linter = new TwigLinter( function( $handle, string $errorMessage ) use ( &$errorWasOutput ) {
+		$linter = new PurifierLinter( function( $handle, string $errorMessage ) use ( &$errorWasOutput ) {
 			$this->assertContains( 'Impure HTML', $errorMessage );
 			$errorWasOutput = true;
 		} );
 
 		$this->expectOutputRegex( '/Validating/' );
-		$linter->lint( __DIR__ . '/data/InvalidTwigFile.twig' );
+		$linter->lint( __DIR__ . '/../data/InvalidHtmlFile.twig' );
 		$this->assertTrue( $errorWasOutput, 'Error should be output' );
 	}
